data_type, value_repr, result, match
